+++ R/scto_write.R
@@ -1,5 +1,8 @@
 #' Write data to a SurveyCTO server
 #'
+#' @description
+#' `r lifecycle::badge('experimental')`
+#'
 #' This function updates an existing dataset using a web POST request, as
 #' uploading data is not officially supported by the SurveyCTO API.
 #'
@@ -23,9 +26,6 @@
 #' auth = scto_auth('scto_auth.txt')
 #' r = scto_write(auth, data, 'my_dataset', 'My Dataset')
 #' }
-#'
-#' @seealso [scto_auth()], [scto_meta()], [scto_read()],
-#'   [scto_get_form_definitions()], [scto_get_attachments()]
 #'
 #' @export
 scto_write = function(