+++ NEWS.md
@@ -1,3 +1,9 @@
+# rsurveycto 0.2.0
+* Added fetching of form metadata, including defintions, for previous and deployed versions using `scto_get_form_metadata()`.
+* Added unnesting of form definitions using `scto_unnest_form_definitions()`.
+* Added information provided by `scto_catalog()`.
+* Fixed `scto_read()` to use and return timestamps in UTC.
+
 # rsurveycto 0.1.6
 * Added parsing of groups.